jwt refactored

+++ src/resources/item.py
@@ -1,12 +1,16 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 class Items(Resource):
-    @jwt_required()
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
-        #return {'items': list(map(lambda i: i.json(), ItemModel.query.all()))} #same funcionality as line above
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {item['name'] for item in ItemModel.find_all()}, 200
+        #return {'items': list(map(lambda i: i.json(), ItemModel.query.all()))} #lambda version to list comprehention
 
 class Item(Resource):
     parser = reqparse.RequestParser()
@@ -20,7 +24,7 @@ class Item(Resource):
         required=True,
         help="Gimme store!"
         )
-
+    @jwt_required()
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -55,6 +59,9 @@ class Item(Resource):
     
     @jwt_required()
     def delete (self, name):
+        claims = get_jwt()
+        if not claims['is admin']:
+            return {'message': 'Admin privilage required'}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
